monitoring/monitorlib/requirements.txt
cryptography==2.9.2
flask==1.1.2
google-auth==1.6.3
jsonschema==3.2.0
jwcrypto==0.7
pyjwt==1.7.1
pyyaml==5.3.1
requests==2.22.0
s2sphere==0.2.5
termcolor==1.1.0
